Refactor EigerDLS formats
Move most of the code to a base class EigerDLS and then subclass this for 4M, 16M as suggested, with the latter having the dynamic masking
